Fix auto TOC generation and minor formatting edit
Fixed a bug where Calibre's automatic structure
detection put Prologue and Epilogue at the
beginning of the book's structure. This was not
part of the in-text TOC, but of the one outside
of the text, generally viewed with EPUB viewers.
Also, changed some things in the markdown to
improve the TOC detection and formatting of the
title.